refactor(mm-next): move getSectionAndTopicFromDefaultHeaderData
