+++ packages/mirror-media-next/pages/external/[slug].js
@@ -15,7 +15,7 @@ import {
   handleAxiosResponse,
   handleGqlResponse,
 } from '../../utils/response-handle'
-import { getSectionAndTopicFromDefaultHeaderData } from '../../utils/api'
+import { getSectionAndTopicFromDefaultHeaderData } from '../../utils/data-process'
 
 /**
  * @typedef {import('../../apollo/fragments/external').External} External
