+++ packages/mirror-media-next/pages/externals/[partnerSlug].js
@@ -26,7 +26,7 @@ import {
   handleAxiosResponse,
   handleGqlResponse,
 } from '../../utils/response-handle'
-import { getSectionAndTopicFromDefaultHeaderData } from '../../utils/api'
+import { getSectionAndTopicFromDefaultHeaderData } from '../../utils/data-process'
 
 const GPTAd = dynamic(() => import('../../components/ads/gpt/gpt-ad'), {
   ssr: false,
