+++ packages/mirror-media-next/pages/marketing/index.js
@@ -1,10 +1,8 @@
 // TODO: add handle-forbid-to-marketing middleware
 
 import styled from 'styled-components'
-import {
-  fetchHeaderDataInDefaultPageLayout,
-  getSectionAndTopicFromDefaultHeaderData,
-} from '../../utils/api'
+import { fetchHeaderDataInDefaultPageLayout } from '../../utils/api'
+import { getSectionAndTopicFromDefaultHeaderData } from '../../utils/data-process'
 import { setPageCache } from '../../utils/cache-setting'
 import Layout from '../../components/shared/layout'
 import BlankCard from '../../components/subscribe/blank-card'
