+++ packages/mirror-media-next/pages/password-change-fail.js
@@ -6,10 +6,8 @@ import { setPageCache } from '../utils/cache-setting'
 import { ENV } from '../config/index.mjs'
 import { getLogTraceObject } from '../utils'
 import { handleAxiosResponse } from '../utils/response-handle'
-import {
-  fetchHeaderDataInDefaultPageLayout,
-  getSectionAndTopicFromDefaultHeaderData,
-} from '../utils/api'
+import { fetchHeaderDataInDefaultPageLayout } from '../utils/api'
+import { getSectionAndTopicFromDefaultHeaderData } from '../utils/data-process'
 
 const Container = styled.div`
   flex-grow: 1;
