+++ packages/mirror-media-next/pages/section/[slug].js
@@ -5,9 +5,9 @@ import SectionArticles from '../../components/shared/section-articles'
 import { ENV } from '../../config/index.mjs'
 import {
   fetchHeaderDataInDefaultPageLayout,
-  getSectionAndTopicFromDefaultHeaderData,
   getPostsAndPostscountFromGqlData,
 } from '../../utils/api'
+import { getSectionAndTopicFromDefaultHeaderData } from '../../utils/data-process'
 import { getLogTraceObject } from '../../utils'
 import {
   handleAxiosResponse,
